Match borica payments by cardId first and then username
Members are creating multiple accounts on bspb.org and simple username is not
enough to import all appropriate payments. CardId is the single unique
identifier that members are given and should be used to primarily identify
payments. That is why this change tries to match payments by cardId first and
if no matching member can be found then it tries using username, which is
provided by administrator.